finish fetch investor profile data and update investor profile data

+++ lib/features/investor_account/data/models/update_investor_profile/update_investor_profile_request.dart
@@ -8,8 +8,8 @@ class UpdateInvestorProfileRequest {
   final String lastName;
   final String investorType;
   final String location;
-  final int minInvestment;
-  final int maxInvestment;
+  final double minInvestment;
+  final double maxInvestment;
   final String stage;
   final String investmentInterests;
   final String bio;
